Split this test up

+++ test/Moose/Maam/Generator/GeneratorTest.php
@@ -6,36 +6,52 @@ use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use Xpmock\TestCase;
 
+/**
+ * Tests the Generator. All tests that invoke the ->generate() method must run in a separate process, as including the
+ * same class file twice will result in a PHP redeclaration error.
+ */
 class GeneratorTest extends TestCase
 {
     protected static $generationPath;
     protected static $assetDir;
 
+    /**
+     * @runInSeparateProcess
+     */
     public function testGenerateGeneratesClassFilesForPhpFilesInTheSourcePath()
     {
         $this->assertTrue(!file_exists(self::$generationPath . '/MaamTest/Person.php'));
-        $this->assertTrue(!file_exists(self::$generationPath . '/NoAnnotationsHere.php'));
         $this->assertTrue(!file_exists(self::$generationPath . '/classmap.php'));
 
         $generator = new Generator(self::$assetDir, self::$generationPath);
-        $generator->generate();
+        $classMap = $generator->generate();
 
         $this->assertSame(
-            file_get_contents(self::$assetDir . '/Person.expected-output.txt'),
+            file_get_contents(self::$assetDir . '/MaamTest/Person.expected-output.txt'),
             file_get_contents(self::$generationPath . '/MaamTest/Person.php')
         );
 
-        $classmap = include self::$generationPath . '/classmap.php';
+        $this->assertSame(
+            "<?php\nreturn " . var_export($classMap, true) . ";",
+            file_get_contents(self::$generationPath . '/classmap.php')
+        );
 
-        $this->assertSame(1, count($classmap));
-
-        $className = array_keys($classmap)[0];
-
-        $this->assertSame('MaamTest\Person', $className);
+        $this->assertArrayHasKey('MaamTest\\Person', $classMap);
         $this->assertSame(
             realpath(self::$generationPath . '/MaamTest/Person.php'),
-            realpath($classmap[$className])
+            realpath($classMap['MaamTest\\Person'])
         );
+    }
+
+    /**
+     * @runInSeparateProcess
+     */
+    public function testDoesNothingForClassesWithNoMaamAnnotations()
+    {
+        $this->assertTrue(!file_exists(self::$generationPath . '/NoAnnotationsHere.php'));
+
+        $generator = new Generator(self::$assetDir, self::$generationPath);
+        $generator->generate();
 
         $this->assertTrue(!file_exists(self::$generationPath . '/NoAnnotationsHere.php'));
     }
@@ -52,7 +68,9 @@ class GeneratorTest extends TestCase
 
     public static function tearDownAfterClass()
     {
-        rmdir(self::$generationPath);
+        if (file_exists(self::$generationPath)) {
+            rmdir(self::$generationPath);
+        }
     }
 
     protected function tearDown()
